% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data.R
\docType{data}
\name{affiliation}
\alias{affiliation}
\title{Department affiliation}
\format{A data.frame with the following columns:
\describe{
\item{year}{integer, fiscal year}
\item{NUMERIC CODE}{integer, department code}
\item{name}{character, last_name first_name}
}}
\usage{
affiliation
}
\description{
Provides the link between individuals (by name) and their department
(by numeric code) from 2012-2019.
}
\keyword{datasets}
