% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizing.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize information in the dataframe.}
\usage{
anonymize(df, cols_to_anon = "name", algo = "crc32")
}
\arguments{
\item{df}{The dataframe you wish to anonymize}

\item{cols_to_anon}{The columns to anonymize. The default is people's names.}

\item{algo}{The algorithms to be used. The available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, and murmur32.}
}
\value{
Returns a dataframe with an anonymized column named 'id'.
}
\description{
Anonymize information in the dataframe.
}
\examples{
DF <- data.frame(name = c("John", "Jon", "Jonathan", "Jon"),
      year = c(2010,2010,2011,2011), pay = c(5000,7000,8000,7000))
anonymize(DF)
df <- anonymize(df=sals18, cols_to_anon = "place_of_residence")
}
