% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_sals.R
\name{anonymize_sals}
\alias{anonymize_sals}
\title{Option to anonymize salary data}
\usage{
anonymize_sals(df, col_to_anon = "name", algo = "crc32")
}
\arguments{
\item{df}{A dataframe that needs to be anonymized}

\item{col_to_anon}{The column name that should be anonymized, usually "name" or other identifying feature}

\item{algo}{The algorithm that will be used to anonymize.}
}
\value{
A dataframe that replaces the identifying column with an anonymized one
}
\description{
Option to anonymize salary data
}
\examples{

}
