% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_Anonymize.R
\name{cy_Anonymize}
\alias{cy_Anonymize}
\title{Anonymize Names}
\usage{
cy_Anonymize(df, col_to_anon = "name", algo = "crc32")
}
\arguments{
\item{df}{The dataframe you wish to anonymize}

\item{col_to_anon}{The column to anonymize. The default 'name'.}

\item{algo}{The algorithms to be used. The available choices are md5 (default), sha1, crc32, sha256, sha512, xxhash32, xxhash64, and murmur32. See the digest R package documentation for details.}
}
\value{
Returns the dataframe with a new anonymized column named 'id'.
}
\description{
Anonymize column of your choice to alpha-numeric code (uses digest package)
}
\examples{
rawdf <- data.frame(name = c("John", "Jon", "Jonathan", "Jon"),
      year = c(2010,2010,2011,2011), pay = c(5000,7000,8000,7000))
cy_Anonymize(rawdf, "name")
anondf <- cy_Anonymize(df=sals18, col_to_anon = "place_of_residence")
}
