% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_GenderStats.R
\name{cy_GenderStats}
\alias{cy_GenderStats}
\title{Function to compare M vs F salaries within a department}
\usage{
cy_GenderStats(data = sals18)
}
\arguments{
\item{data}{A dataframe of ISU salary data with academic department info. Default is for year 2018. Column names must include 'department', 'position', 'gender', and 'total_salary_paid'.}
}
\value{
A dataframe of department, nested data, p-value for gender pay gap after accounting for position, and a verdict
}
\description{
Function to compare M vs F salaries within a department
}
\examples{
data(sals18)
sals18 \%>\%
dplyr::filter(department == "AGRONOMY") \%>\%
dplyr::filter(stringr::str_detect(position, "PROF")) \%>\%
cy_GenderStats()

}
