% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{cy_RunShiny}
\alias{cy_RunShiny}
\title{Run Shiny App}
\usage{
cy_RunShiny()
}
\value{
a Shiny App
}
\description{
Launch Shiny App that displays pay information for subsets of Iowa State University Employees
}
\examples{
\dontrun{
cy_RunShiny()
}
}
