% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cy_SimpProfs.R
\name{cy_SimpProfs}
\alias{cy_SimpProfs}
\title{cy_SimpProfs}
\usage{
cy_SimpProfs(data)
}
\arguments{
\item{data}{A dataframe of salary data with a variable 'position'}
}
\value{
profs A dataframe of professors' salary data with tidied position
categories in a new column titled 'pos_simp'
}
\description{
Create a dataframe that only contains professors' salary data
from a larger dataframe of salary data. Reassign position as a simplified title in one of four #'   categories: assistant, associate, full professor, and other.
}
\details{
Extract professors' salary data and assign simplified categories of 'professor'
}
\examples{
cy_SimpProfs()

}
