% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyd_gendept}
\alias{cyd_gendept}
\title{Male and female representation in departments by fiscal year}
\format{A data frame with 952 rows and 6 variables:
\describe{
\item{fiscal_year}{fiscal year as defined by the state of Iowa (July 1 - June 30), currently includes 2007-2018}
\item{college}{Iowa State University college the department is housed in}
\item{dept}{Iowa State University department}
\item{M}{number of male tenure-track professors}
\item{F}{number of female tenure-track professors}
\item{tot}{total number of tenure-track professors}
\item{fracF}{fraction of total professors that are female}
...
}}
\usage{
cyd_gendept
}
\description{
Male and female representation in departments by fiscal year
}
\keyword{datasets}
