% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyd_saldept}
\alias{cyd_saldept}
\title{Salaries of Iowa State University employees with 'prof' in their position title}
\format{A data frame with 16570 rows and 9 variables:
\describe{
\item{fiscal_year}{fiscal year as defined by the state of Iowa (July 1 - June 30), currently includes 2007-2018}
\item{gender}{M (male), F (female), all other categories have been removed in this dataset}
\item{position}{all position titles containing the string 'prof', currently has 36 categories}
\item{total_salary_paid}{total salary, to the penny, is not simply base plus travel}
\item{travel_subsistence}{travel allowances paid, to the penny}
\item{base_salary}{base salary paid, to the penny}
\item{dept}{Iowa State University department the person is housed in}
\item{college}{Iowa State University college the department is housed in}
\item{id}{an anonymized code for each individual}
...
}}
\source{
\url{https://data.iowa.gov/}
}
\usage{
cyd_saldept
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
