% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyd_salprofs}
\alias{cyd_salprofs}
\title{Salaries of Iowa State University employees in one of the six tenured professor categories}
\format{A data frame with 13453 rows and 10 variables:
\describe{
\item{fiscal_year}{fiscal year as defined by the state of Iowa (July 1 - June 30), currently includes 2007-2018}
\item{gender}{M (male), F (female), all other categories have been removed in this dataset}
\item{position}{position titles in one of 6 categories including
\item{total_salary_paid}{total salary, to the penny, is not simply base plus travel}
\item{travel_subsistence}{travel allowances paid, to the penny}
\item{base_salary}{base salary paid, to the penny}
\item{dept}{Iowa State University department the person is housed in}
\item{college}{Iowa State University college the department is housed in}
\item{id}{an anonymized code for each individual}
\item{prof_simp}{further simplified position categories including assoc prof, asst prof, awarded prof, and prof}
...
}}
\source{
\url{https://data.iowa.gov/}
}
\usage{
cyd_salprofs
}
\description{
cyd_saldept, but filtered to only contain on of six tenured categories (assoc prof, asst prof, distg prof, morrill prof, prof, univ prof). Department heads were also eliminated for this dataset. Around 3000 entries were removed during this filtering.
}
\keyword{datasets}
