% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyf_getsals.R
\name{cyf_getsals}
\alias{cyf_getsals}
\title{Get Iowa State University salaries}
\source{
\url{https://db.desmoinesregister.com/state-salaries-for-iowa/}
}
\usage{
cyf_getsals(token, limit = 2e+05, offset = 0)
}
\arguments{
\item{token}{API token}

\item{limit}{max number of entries}

\item{offset}{where to start gathering (useful if you need to get the data in batches)}
}
\value{
A data.frame containing the following columns:
\describe{
\item{fiscal_year}{numeric, fiscal year for this observation}
\item{name}{character, individual's name}
\item{gender}{character, Male, Female, or *}
\item{position}{character, position}
\item{base_salary_date}{Date}
\item{total_salary_paid}{numeric, total salary including bonuses and summer salary}
\item{travel_subsistence}{numeric, payment for travel}
\item{base_salary}{numeric, base salary}
}
}
\description{
This function will obtain Iowa State University salaries from the Des Moines
Register data base. It uses an API token which you will need to request from
here.
}
\details{
Each row contains a person-year.
}
