% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data.R
\docType{data}
\name{departments}
\alias{departments}
\title{Departments}
\format{A data.frame with 261 rows and the following columns:
\describe{
\item{NUMERIC CODE}{integer, numeric code for department}
\item{ALPHA CODE}{character, alpha code for department}
\item{DIRECTORY NAME}{character, department name as used in the directory}
\item{FULL NAME}{character, full department name}
\item{ADDRESS}{character, address}
\item{PHONE}{character, phone number}
\item{RMM RESOURCE UNIT NUMBER}{integer, ??}
\item{PARENT DEPT NUMBER}{integer, code for parental department}
}}
\source{
\url{https://www.ir.iastate.edu/deptcodes}
}
\usage{
departments
}
\description{
Information on departments including name, codes, address, phone number, etc.
}
\keyword{datasets}
