% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_academic_pipeline.R
\name{get_academic_pipeline}
\alias{get_academic_pipeline}
\title{get_academic_pipeline}
\usage{
get_academic_pipeline(dataframe)
}
\arguments{
\item{dataframe}{A dataframe of salary data with a variable 'position'}
}
\value{
academic_pipeline A dataframe of professors' salary data with tidied position
categories
}
\description{
Create a dataframe that contains salaray data
for graduate assistants, lecturers, and professors
}
\details{
Create a dataframe of salary data for positions in the academic pipeline
}
\examples{
get_academic_pipeline(all_sals)

}
