% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profs.R
\name{get_profs}
\alias{get_profs}
\title{get_profs}
\usage{
get_profs(dataframe)
}
\arguments{
\item{dataframe}{A dataframe of salary data with a variable 'position'}
}
\value{
profs A dataframe of professors' salary data with tidied position
categories
}
\description{
Create a dataframe that only contains professors' salaray data
from a larger dataframe of salary data. Create a new variable 'position_simplified'
that groups professor titles into categories such as associate, visting,
and emeritus.
}
\details{
Create a dataframe of professors' salary data
}
\examples{
get_profs(dataframe)

}
