% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_df.R
\name{sal_df}
\alias{sal_df}
\title{Create a dataframe of salary data}
\usage{
sal_df(limit = 1000, offset = 0, fiscal_year = 2007, token = NULL)
}
\arguments{
\item{limit}{How many data entries you'd like to receive (default = 1000)}

\item{offset}{Where you'd like the dataentries to start pulling from (default = 0)}

\item{fiscal_year}{The fiscal year the data are taken from. Limited to 2007-2018}

\item{token}{An API token. Only necessary for large amounts of datascraping. Generated from this \href{"https://dev.socrata.com/foundry/data.iowa.gov/s3p7-wy6w"}{website}}
}
\value{
A dataframe with salary information, position, and date for Iowa State University employees
}
\description{
Scrape ISU salary data by year from the web and turn it into a tidy dataframe

See \href{https://github.com/vanichols/CyChecks}{CyChecks}
}
\details{
An API (or APP) token isn't necessary for scraping data, but it will help speed up the data grabbing process and will allow users to get nearly unlimited data.
}
\examples{
\dontrun{
sal_df()
x <- sal_df(limit = 10, fiscal_year = 2015)
}
}
\author{
Lydia English
}
