% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Salaries}
\format{A data.frame with the following columns:
\describe{
\item{year}{integer, fiscal year for this salary}
\item{name}{character, individual's name}
\item{gender}{character, \link{M}ale, \link{F}emale, or \link{*}}
\item{place_of_residence}{character, county}
\item{position}{character, name of position}
\item{base_salary_date}{date-time, it seems to always be 1 July of that fiscal_year}
\item{total_salary_paid}{numeric, total salary including summer pay, bonuses etc}
\item{travel_subsistence}{numeric, reimbursement (?) for travel}
\item{base_salary}{numeric, base pay}
\item{pay_period}{character, 'hourly' or 'yearly'}
}}
\source{
\url{https://db.desmoinesregister.com/state-salaries-for-iowa}
}
\usage{
salaries
}
\description{
Iowa State University salaries as downloaded from the Des Moines Register.
Base_salary has been cleaned to deal with hourly pay, hyphens, and commas.
A new column \code{pay_period} has been introduced to distinguish hourly pay from
yearly pay. No differentiation is made concerning A-base (12-month) and
B-base (9-month) pay in the database.
}
\keyword{datasets}
