% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sals18.R
\docType{data}
\name{sals18}
\alias{sals18}
\title{Salary information for Iowa State Univeristy employees in 2018}
\format{A dataframe with 9,989 rows and 10 variables
\describe{
\item{base_salary_date}{the date base salary was computed}
\item{fiscal_year}{year (July 1 - June 30) salary data is associated with}
\item{gender}
\item{name}{name of employee}
\item{place_of_residence}{county or city whereby employee resides}
\item{position}{position within the university}
\item{total_salary_paid}{total amount (in dollars) of salary for that fiscal year}
\item{travel_subsistence}{amount (in dollars) of travel expenditures reimbursed during the fiscal year}
\item{organization}{larger umbrella organization under which the employee works}
\item{department}{home department on campus}
}}
\source{
\url{https://data.iowa.gov/State-Finances/State-of-Iowa-Salary-Book/s3p7-wy6w}
}
\usage{
sals18
}
\description{
A dataset containing the salary information, position, and gender of over
9,000 employees at Iowa State University
}
\keyword{datasets}
