% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_profs.R
\name{simplify_profs}
\alias{simplify_profs}
\title{Simplify professor titles}
\usage{
simplify_profs(data = cyd_saldept)
}
\arguments{
\item{data}{A dataframe that can be simplified.}
}
\value{
A dataframe without adjunct, affiliation, emeritus, visiting, collaborating or department chairs,
as their salaries tend to be weird.
}
\description{
Simplify professor titles
}
\examples{


}
