% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_by1.R
\name{slice_by1}
\alias{slice_by1}
\title{slice_by1}
\usage{
slice_by1(all_sals,slice_variable)
}
\arguments{
\item{slice_var}{A variable to group_by}

\item{all_sals}{A dataframe of ISU salary data. Default is for year 2007.}
}
\value{
A dataframe of mean, median, min, 1st quartile, 3rd quartile, and max of total_salary_paid, grouped by slice_variable
}
\description{
Function to slice dataframe of salaries by a single variable
}
\examples{
slice_by1(all_sals,'gender')
slice_by(all_sals,'position')
}
