% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_by2.R
\name{slice_by2}
\alias{slice_by2}
\title{slice_by2}
\usage{
slice_by2(all_sals,slice_var1,slice_var2)
}
\arguments{
\item{slice_var1}{First variable to group_by}

\item{slice_var2}{Second variable to group_by}

\item{all_sals}{A dataframe of ISU salary data. Default is for year 2007.}
}
\value{
A dataframe of mean, median, min, 1st quartile, 3rd quartile, and max of total_salary_paid, grouped by slice_var1 and slice_var2
}
\description{
Function to slice dataframe of salaries by two variables
}
\examples{
slice_by1(all_sals,'gender','position')
}
