% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizing.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize some information in the dataframe.}
\usage{
anonymize(df, cols_to_anon = "name", algo = "crc32")
}
\arguments{
\item{df}{The name of the file includes a dataframe.}

\item{cols_to_anon}{The columns to anonymize. The default is people's names}

\item{algo}{The algorithms to be used. The available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, and murmur32.}
}
\value{
Returns a dataframe with an anonymized column (id).
}
\description{
Anonymize some information in the dataframe.
}
\examples{
DF <- sal_df()
anonymize(DF)
}
