% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_df.R
\name{sal_df}
\alias{sal_df}
\title{Create a dataframe of salary data}
\usage{
sal_df(limit = 1000, offset = 0, fiscal_year = 2007, token = NULL)
}
\arguments{
\item{limit}{The number of data entries (rows in the dataframe) you'd like to receive (default = 1000)}

\item{offset}{Where you'd like the data entries to start pulling from (default = 0). If using the
default of 0, then the dataset will start pulling from the beginning of alphabet.
This argument is useful to change when using the function multiple times.}

\item{fiscal_year}{The fiscal year the data are taken from. Limited to 2007-2018.
Can only enter one fiscal year at a time.}

\item{token}{An API token. Only necessary for large amounts of datascraping. Generated from this \href{"https://dev.socrata.com/foundry/data.iowa.gov/s3p7-wy6w"}{website}}
}
\value{
A dataframe with salary information, position, and gender for Iowa State University employees in a given fiscal year.
}
\description{
Scrape ISU salary data by year from the web and turn it into a tidy dataframe

See \href{https://github.com/vanichols/CyChecks}{CyChecks}
}
\details{
An API (or APP) token isn't necessary for scraping data, but it will help speed up the data grabbing process and will allow users to get nearly unlimited data.
}
\examples{
sal_df()
x <- sal_df(limit = 10, fiscal_year = 2015)
}
\author{
Lydia English
}
