% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mf.R
\name{stats_mf}
\alias{stats_mf}
\title{Function to compare M vs F salaries within a department}
\usage{
stats_mf(data = sals18)
}
\arguments{
\item{data}{A dataframe of ISU salary data with academic department info. Default is for year 2018.}
}
\value{
A dataframe of department, nested data, p-value for gender pay gap, and a verdict
}
\description{
Function to compare M vs F salaries within a department
}
\examples{
stats_mf(data = filter(sals18, department == "AGRONOMY", grepl("PROF", position)))
}
